requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiosqlite>=0.19.0
httpx>=0.25.0
pyyaml>=6.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
